251031/Tarasenko/lab 4/publisher/requirements.txt
Flask==2.2.2
kafka-python==2.0.2
dataclasses; python_version < '3.7'
